LayoutTests/http/tests/canvas/philip/tests/security.pattern.image.fillStyle.html

<!DOCTYPE html>
<title>Canvas test: security.pattern.image.fillStyle</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="security.pattern.canvas.timing.html" title="security.pattern.canvas.timing">
<link rel="next" href="security.pattern.canvas.fillStyle.html" title="security.pattern.canvas.fillStyle">
<body class="show_output">
<p>
 <a href="security.pattern.canvas.timing.html" accesskey="p" title="[p] security.pattern.canvas.timing">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="security.pattern.canvas.fillStyle.html" accesskey="n" title="[n] security.pattern.canvas.fillStyle">&gt;</a>
<h1><a href="index.security.html">security</a>.<a href="index.security.pattern.html">pattern</a>.<a href="index.security.pattern.image.html">image</a>.fillStyle</h1>
<p class="desc">Setting fillStyle to a pattern of a different-origin image makes the canvas origin-unclean</p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.security.fillStyle.image">security.fillStyle.image</a>

</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

var p = ctx.createPattern(document.getElementById('yellow.png'), 'repeat');
ctx.fillStyle = p;
ctx.fillStyle = 'red';
try { var _thrown = false; canvas.toDataURL(); } catch (e) { _thrown = true; } finally { _assert(_thrown, "should throw exception: canvas.toDataURL()"); }
try { var _thrown = false; ctx.getImageData(0, 0, 1, 1); } catch (e) { _thrown = true; } finally { _assert(_thrown, "should throw exception: ctx.getImageData(0, 0, 1, 1)"); }


});
</script>
<img src="http://localhost:8000/canvas/philip/images/yellow.png" id="yellow.png" class="resource">
