LayoutTests/http/tests/canvas/philip/tests/security.pattern.image.fillStyle-expected.txt

CONSOLE MESSAGE: line 32: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
< [index] >
security.pattern.image.fillStyle
Setting fillStyle to a pattern of a different-origin image makes the canvas origin-unclean
References: security.fillStyle.image
Actual output:
Passed
